初始化了PRPC模块的服务发现注册模块

+++ common/utils/utills.go
@@ -19,5 +19,5 @@ func HashStr(key string) uint32 {
 }
 
 func HashWithSlot(hash uint32, slot int) uint32 {
-	return hash % slot
+	return hash % uint32(slot)
 }
